Alter default scope separator for Xero connect requests
The default oauth league scope separator is a comma, however Xero
requires spaces instead. When passing an array of scopes, the underlying
league code will implode the scope array automatically. But this will be
rejected via Xero due to the comma. Use a space instead.

@see https://github.com/calcinai/oauth2-xero#authorization-code-flow
@see https://developer.xero.com/documentation/oauth2/sign-in

Scopes are the permissions that your app is requesting approval for.
You can pass in a list of scopes separated by a space.

Remove extra space, new line